SEM1-241-long-extensions-unit-tests (#6)
* 1st draft

* Done for now(Randomize.Integers.Long(min, max)) needs a look at ( still returns negative values for 0 set as min).

* 2nd draft

* Update GetMostSignificantBit as it was not working as intended

* Update GetMostSignificantBit test as the test was incorrect

* Update Theory Test for GetMostSignificant Bit to reflect changes and fix unit tests

* Revert Big Integer tests as it is done in separate pr.

* Long GetMostSignificantBit adjust

+++ Source/Emvco/Core/Play.Core/Extensions/Integers/UIntExtension.cs
@@ -11,7 +11,7 @@ public static class UIntExtension
     public static bool AreAnyBitsSet(this uint value, uint bitsToCompare) => (value & bitsToCompare) != 0;
     public static uint ClearBits(this in uint input, uint bitsToClear) => input & ~bitsToClear;
     public static string GetBinaryString(this uint value) => Convert.ToString(value, 2);
-    public static uint GetMaskedValue(this uint value, in uint mask) => (byte) (value & ~mask);
+    public static uint GetMaskedValue(this uint value, in uint mask) => value & ~mask;
 
     public static int GetMostSignificantBit(this uint value)
     {
